Move code into ReaderThemeSupportFeatures service
