+++ src/AmpWpPlugin.php
@@ -99,6 +99,7 @@ final class AmpWpPlugin extends ServiceBasedPlugin {
 		'plugin_registry'                    => PluginRegistry::class,
 		'plugin_suppression'                 => PluginSuppression::class,
 		'reader_theme_loader'                => ReaderThemeLoader::class,
+		'reader_theme_support_features'      => ReaderThemeSupportFeatures::class,
 		'rest.options_controller'            => OptionsRESTController::class,
 		'rest.validation_counts_controller'  => Validation\ValidationCountsRestController::class,
 		'server_timing'                      => Instrumentation\ServerTiming::class,
@@ -193,6 +194,7 @@ final class AmpWpPlugin extends ServiceBasedPlugin {
 			DevTools\CallbackReflection::class,
 			DevTools\FileReflection::class,
 			ReaderThemeLoader::class,
+			ReaderThemeSupportFeatures::class,
 			BackgroundTask\BackgroundTaskDeactivator::class,
 			PairedRouting::class,
 			Injector::class,
